moved all of the Frame code to its own file. finished porting all of the cluster update code into the Frame module. changed all of my testing and playground code to work with the new framework. did a plotting test for the 2d system with the new version and everything seems to work the same as before.
