+++ src/body/cluster.py
@@ -27,416 +27,14 @@ import warnings
 import sys
 import os
 
-from enum import Enum
-
 from body import neighborgrid as ng
 from body import body as body
-
-class Event(Enum):
-
-    NEW_CLUSTER    = 0
-    POSSIBLE_SPLIT = 1
-    MERGE          = 2
-    SPLIT          = 3
-    PERSIST        = 4
-
-
-class Frame:
-
-    def __init__(self, bodies, clusters, frame_num, monomer_frac = -1):
-
-        #init a dictionary that stores which id gets mapped to which cluster
-        self.__label_dict = dict()
-
-        #init a dictionary that stores which old clusters update to new ones
-        self.__updates    = dict()
-
-        #init the clusters and bodies lists
-        self.__clusters   = clusters
-        self.__bodies     = bodies
-
-        #init the other values
-        self.__monomer_frac = monomer_frac
-        self.__frame_num    = frame_num
-
-
-    def create_first_frame(self, cluster_info, frame_num, observer):
-        #for first frame, just create cluster info for all existing clusters
-
-        for orig_cluster in self.__clusters:
-
-            self.__add_new_cluster(cluster_info, orig_cluster, frame_num,  
-                                   self.__monomer_frac, observer)
-
-            #do computations inolving monomers
-            num_bodies = len(orig_cluster.get_bodies())
-            cluster_info[-1].add_monomers(orig_cluster, num_bodies, 1)
-
-        return
-
-    def get_similarity(self, old_cluster, new_cluster):
-        #determine a similarity measure between old and new clusters
-        #also return an event for split or persist
-
-        #perform a set difference to determine gain/loss of subunits
-        old_not_new = old_cluster.get_difference(new_cluster)
-        new_not_old = new_cluster.get_difference(old_cluster)
-
-        #define a similarity, max of the two set differences
-        similarity = max(len(old_not_new), len(new_not_old))
-
-        #determine if the cluster has split into a potentially new cluster or not
-        event = Event.PERSIST
-        if len(old_not_new) > 1:    
-            event = Event.SPLIT
-
-        return similarity, event
-
-    def __assign_match(self, cluster, match, similarity):
-        #assign cluster to its match in the previous frame. update relevant dict entries
-
-        #grab the cluster id and set it for the new cluster
-        cluster_id = match.get_cluster_id()
-        cluster.set_cluster_id(cluster_id)
-
-        #update the update and label dicts
-        self.__updates[match] = cluster
-        self.__label_dict[cluster_id] = [similarity, cluster]
-
-        return
-
-    def __assign_better_match(self, cluster, match, similarity):
-        #cluster is closer to match than theft victim. steal ids and update dicts
-
-        match_id = match.get_cluster_id()
-
-        #current cluster is more similar to match. overwrite old cluster
-        theft_victim = self.__label_dict[match_id][1]
-        cluster.steal_id(theft_victim)
-
-        #overwrite entries to the update and label dicts
-        self.__updates[match] = cluster
-        self.__label_dict[match_id] = [similarity, cluster]
-
-        #return the theft victim
-        return theft_victim
-
-
-    def __handle_split(self, match, cluster, similarity, cluster_info, observer):
-        '''
-        If a cluster splits, determine the best match as the one most similar to the
-        original. It is possible for a match to be assigned, but a better one to be found 
-        later, in which case we steal the ID from the previous match and return a flag
-        to add the old match back to the queue.
-        '''
-
-        #set frame num and monomer frac
-        frame_num = self.get_frame_num()
-        current_monomer = self.get_monomer_fraction()
-
-        #get the id for the matched cluster
-        match_id = match.get_cluster_id()
-
-        #check if this match has been assigned already
-        if match_id in self.__label_dict:
-
-            #check if the current similarity score is better 
-            if similarity < self.__label_dict[match_id][0]:
-
-                theft_victim = self.__assign_better_match(cluster, match, similarity)
-                # print("Better match found. Updated cluster match to ", match_id)
-
-                #return the old cluster for reassignment to the queue
-                return theft_victim
-
-            #if similarity is worse, assign a new cluster
-            else: 
-
-                self.__add_new_cluster(cluster_info, cluster, frame_num,  
-                                       self.__monomer_frac, observer)
-
-                #when splitting, set the parent to be the matched cluster
-                cluster_info[-1].set_parent(match)
-                # print("Match found with worse similarity. Created new cluster with id ", cluster_num)
-
-        #this matching old cluster has not been assigned, so assign it
-        else:
-
-            self.__assign_match(cluster, match, similarity)
-            # print("matched cluster to existing ", cluster_id)
-
-
-        return None
-
-
-
-
-
-    def update(self, cluster_info, old_frame, observer):
-        #use the previous frame data to update the new frame
-
-        #make a queue for the new clusters and dict to store the mappings
-        queue = self.__clusters.copy() 
-
-        #make a list to store death updates for merging - entries (dying id, new cluster id)
-        merge_updates = []
-
-        #get required data from current and previous frame
-        prev_monomer = old_frame.get_monomer_fraction()
-        old_bodies   = old_frame.get_bodies()
-        current_monomer = self.get_monomer_fraction()
-        frame_num       = self.get_frame_num()
-        bodies          = self.get_bodies()
-
-
-        #loop over the queue to assign labels to each new cluster
-        while len(queue) > 0:
-
-            #grab the first cluster in queue
-            cluster = queue.pop(0)
-
-            #determine which previous clusters are possible matches for the current cluster
-            possibleMatches = cluster.get_possible_matches(old_bodies)
-
-            #get an event based on the possible matches
-            event = self.__determine_event(possibleMatches)
-            print(event)
-
-            if (event == Event.NEW_CLUSTER):
-                #create a new cluster and append to cluster_info
-
-                self.__add_new_cluster(cluster_info, cluster, frame_num,  
-                                       current_monomer, observer)
-
-                #augment with monomer addition info
-                num_bodies = len(cluster.get_bodies())
-                cluster_info[-1].add_monomers(cluster, num_bodies, 1)
-
-            elif (event == Event.POSSIBLE_SPLIT):
-                #clusters have possibly split. perform matching if they have
-               
-                #extract the matching cluster from possibilities list
-                match = list(possibleMatches)[0]
-
-                #determine similarity between match and cluster. determine if split possible
-                similarity, sub_event = self.get_similarity(match, cluster)
-                print(sub_event)
-
-                if (sub_event == Event.SPLIT):
-                    #if two or more particles split from the cluster. Find best match
-
-                    to_queue = self.__handle_split(match, cluster, similarity, 
-                                                   cluster_info, observer)
-                    if (to_queue is not None):
-                        queue.append(to_queue)
-
-
-                #if the difference is 1 or 0, same cluster or only lost one monomer
-                elif (sub_event == Event.PERSIST):
-
-                    #assign match to cluster
-                    self.__assign_match(cluster, match, similarity)
-                    
-                    # print("matched cluster to existing ", cluster_id)
-
-            elif (event == Event.MERGE):
-
-                
-                # print("Clusters have merged")
-
-                #get a list of possible matches, and assign a similarity to each
-                possible_list   = list(possibleMatches)
-                similarity_vals = []
-                for possible_match in possible_list:
-
-                    #do a set difference between cluster and possible
-                    old_not_new = set(possible_match.get_body_ids()).difference(set(cluster.get_body_ids()))
-                    new_not_old = set(cluster.get_body_ids()).difference(set(possible_match.get_body_ids()))
-                    similarity = max(len(old_not_new), len(new_not_old))
-                    # similarity = len(new_not_old)
-                    # print("Similarity (Merge): ", similarity)
-
-                    similarity_vals.append(similarity)
-
-                #sort the possibilities by similarity
-                poss_aug = [(similarity_vals[i], i) for i in range(len(similarity_vals))]
-                poss_aug.sort()
-                sorted_sim,permutation = zip(*poss_aug)
-                sorted_poss = [possible_list[i] for i in permutation]
-
-                #loop through sorted list. assign this cluster the first un-assigned match
-                match_flag = False
-                for possibility in sorted_poss:
-
-                    #check if match not already assigned. If not, assign it. 
-                    match_id = possibility.get_cluster_id()
-                    if (not match_flag):
-                        if (match_id not in self.__label_dict): #removed .keys() in case things break
-
-                            # possibility.update(cluster)
-                            self.__updates[possibility] = cluster
-
-                            #set id for the new cluster
-                            cluster.set_cluster_id(match_id)
-                            self.__label_dict[match_id] = [similarity, cluster]
-                            match_flag = True
-                            assigned_id = match_id
-
-                        else:
-                            continue
-
-                    #match is already assigned, set other possibilities to update to it and die 
-                    else:
-
-                        merge_updates.append((match_id, assigned_id))
-
-                #if no match has been made after all possibilities, create new cluster
-                if not match_flag:
-
-                    #i dont think its possible to get here, but we shall see
-                    # print("All clusters assigned. Creating new cluster")
-
-                    #get the new cluster id
-                    cluster_num = len(cluster_info)
-
-                    #update the cluster with the new id, add entry to cluster_info
-                    cluster.set_cluster_id(cluster_num)
-                    self.__label_dict[cluster_num] = [0, cluster]
-                    cluster_info.append(ClusterInfo(cluster, frame, observer))
-                    # print("created new (merge) cluster with id ", cluster_num)
-
-
-        #loop over the tentative updates and apply them. do monomer checking first
-        for key in self.__updates.keys():
-
-            #get lists of the body ids in the pre and post update clusters
-            past_ids = set(key.get_body_ids())
-            new_ids  = set(self.__updates[key].get_body_ids())
-
-            #determine how many monomers were gained and lost from these id sets
-            m_gain, m_lost = get_monomer_stats(past_ids, new_ids, old_bodies, bodies)
-            # print("Monomer gain {}, Monomer loss {}".format(m_gain, m_lost))
-
-            #update the monomer stats
-            if m_gain > 0:
-
-                cluster_id = key.get_cluster_id()
-                monomer_frac = observer.previous_monomer
-                cluster_info[cluster_id].add_monomers(self.__updates[key], m_gain, monomer_frac)
-
-            if m_lost > 0:
-
-                cluster_id = key.get_cluster_id()
-                cluster_info[cluster_id].remove_monomers(key, m_lost)
-
-
-            #perform the update on the cluster
-            key.update(self.__updates[key])
-
-        #loop over the now matched clusters and update the cluster_info
-        #(can loop over all keys in the label_dict to avoid unnec work)
-        for key in list(self.__label_dict.keys()):
-
-            #get the cluster
-            updated_cluster = self.__label_dict[key][1]
-
-            #update the clusterinfo
-            cluster_info[key].update_data(updated_cluster, frame_num, observer.current_monomer)
-            # print("Updated cluster {}".format(key))
-
-        #loop over the clusters that merged and are no longer tracked
-        for update_pair in merge_updates:
-
-            old_id = update_pair[0]
-            current_id = update_pair[1]
-            current_cluster = self.__label_dict[current_id][1]
-
-            #update to cluster, but also set a death
-            cluster_info[old_id].update_data(current_cluster, frame_num, observer.current_monomer)
-            cluster_info[old_id].kill(frame_num)
-            # print("Updated and killed cluster {} (Merge)".format(old_id))
-
-
-        return cluster_info
-
-
-
-
-    def get_frame_num(self):
-
-        return self.__frame_num
-
-    def get_bodies(self):
-
-        return self.__bodies
-
-    def get_clusters(self):
-
-        return self.__clusters
-
-    def get_monomer_fraction(self):
-
-        return self.__monomer_frac
-
-    def __determine_event(self, possibleMatches):
-        #return what kind of update to do based on length of possible matches
-
-        L = len(possibleMatches)
-
-        if L == 0:
-
-            '''
-            If there are no possibleMatches, none of the bodies in this cluster were 
-            in clusters during the previous timestep. Thus this is a new cluster. 
-            '''
-
-            return Event.NEW_CLUSTER
-
-        elif L == 1:
-
-            '''
-            If there is exactly one possible match, then all the bodies in this cluster were 
-            either in the same cluster, or not part of a cluster during the previous timestep. 
-            Several cases are possible.
-            1) The structure is exactly the same. 
-            2) The structure has gained monomers 
-            3) The structure has broken into sub-structures
-            '''
-
-            return Event.POSSIBLE_SPLIT
-
-        else:
-
-            '''
-            If there is more than one potential match, that means that two or more clusters 
-            from the previous timestep have merged during this frame. 
-
-            May also be able to get here if clusters merge and break at the same time. 
-            '''
-
-            return Event.MERGE
-
-
-    def __add_new_cluster(self, cluster_info, cluster, frame_num, current_monomer, observer):
-        #create new cluster and append it to cluster info
-
-        #get the id number for the new cluster by looking at length of cluster info
-        cluster_num = len(cluster_info)
-
-        #update the cluster with the new id, add entry to cluster_info
-        cluster.set_cluster_id(cluster_num)
-        self.__label_dict[cluster_num] = [0, cluster]
-        cluster_info.append(ClusterInfo(cluster, frame_num, current_monomer, observer))
-
-        # print("created new cluster with id ", cluster_num)
-
-        return
-
+from body import frame as frame
 
 
 class Cluster:
 
-    def __init__(self, bodies, frame):
+    def __init__(self, bodies, frame_num):
 
         #create a reference to the list of bodies comprising the cluster
         self.__bodies = bodies
@@ -445,7 +43,7 @@ class Cluster:
         self.__cluster_index = -1
 
         #set a last updated value
-        self.__last_updated = frame
+        self.__last_updated = frame_num
 
 
     def get_possible_matches(self, old_bodies):
@@ -577,19 +175,19 @@ class ClusterInfo:
 
         return
 
-    def kill(self, frame):
+    def kill(self, frame_num):
         # set this cluster to dead status
 
-        self.__death_frame = frame
+        self.__death_frame = frame_num
         self.__set_lifetime()
         self.__is_dead = True
 
         return
 
-    def update_data(self, cluster, frame, monomer_frac):
+    def update_data(self, cluster, frame_num, monomer_frac):
         #append the current cluster's coordinate data to storage
 
-        if (frame > self.__last_updated and not self.__is_dead):
+        if (frame_num > self.__last_updated and not self.__is_dead):
 
             #compute a dictionary of requested values using this cluster
             self.__stored_data.append(self.__compute_coordinate(cluster))
@@ -598,7 +196,7 @@ class ClusterInfo:
             self.__stored_data[-1]['monomer_fraction'] = monomer_frac
 
             #updated the time of last update to current frame
-            self.__last_updated = frame
+            self.__last_updated = frame_num
 
         return
 
@@ -716,13 +314,6 @@ class Observer:
         self.__observable_set = set(['positions'])
 
 
-
-
-
-
-
-
-
 ####################################################################
 ################# Bond Dict -> Graph & Clustering ##################
 ####################################################################
@@ -813,355 +404,15 @@ def get_group_sizes(G):
 
 
 
-def get_monomer_stats(old_ids, new_ids, old_bodies, new_bodies):
-    #determine the number of monomer additions and subtractions from old to new cluster
-
-    #init counters for number of monomers gained and lost
-    monomers_gained = 0
-    monomers_lost   = 0
-
-    #first, do additions. look at ids in new but not old, and check cluster in old_bodies
-    new_not_old = new_ids.difference(old_ids)
-    for entry in new_not_old:
-
-        cluster_id = old_bodies[entry].get_cluster_id()
-        if cluster_id == -1:
-            monomers_gained += 1
-
-    #next do subtractions. look at ids in old but not new. check cluster in bodies
-    old_not_new = old_ids.difference(new_ids)
-    for entry in old_not_new:
-
-        cluster_id = new_bodies[entry].get_cluster_id()
-        if cluster_id == -1:
-            monomers_lost += 1
-
-    return monomers_gained, monomers_lost
-
-
-
-def update_clusters(clusters, cluster_info, bodies, old_bodies, frame, observer):
-    #use the existing clusters to assign an id and update cluster
-
-    #make a queue for the new clusters and dict to store the mappings
-    queue = [cluster for cluster in clusters] #TODO change to list.copy()
-    label_dict = dict()
-
-    #make a list to store death updates for merging - entries (dying id, new cluster id)
-    merge_updates = []
-
-    #make a dict to store tentative updates during the matching. Apply at the end
-    tentative_updates = dict()
-
-    #loop over the queue to assign labels to each new cluster
-    while len(queue) > 0:
-
-        #grab the first cluster in queue
-        cluster = queue.pop(0)
-
-        #determine which previous clusters are possible matches for the current cluster
-        possibleMatches = get_possible_matches(cluster, old_bodies)
-
-        if len(possibleMatches) == 0:
-
-            '''
-            If there are no possibleMatches, none of the bodies in this cluster were in clusters
-            during the previous timestep. Thus this is a new cluster. Create a new cluster whose
-            id is the number of elements in cluster_info.
-            '''
-
-            #get the new cluster id
-            cluster_num = len(cluster_info)
-
-            #update the cluster with the new id, add entry to cluster_info
-            cluster.set_cluster_id(cluster_num)
-            label_dict[cluster_num] = [0, cluster]
-            cluster_info.append(ClusterInfo(cluster, frame, observer))
-
-            num_bodies = len(cluster.get_bodies())
-            monomer_frac = observer.previous_monomer
-            cluster_info[cluster_num].add_monomers(cluster, num_bodies, monomer_frac)
-            # print("created new cluster with id ", cluster_num)
-
-        elif len(possibleMatches) == 1:
-
-            '''
-            If there is exactly one possible match, then all the bodies in this cluster were 
-            either in the same cluster, or not part of a cluster during the previous timestep. 
-            Several cases are possible.
-            1) The structure is exactly the same. 
-            2) The structure has gained monomers 
-            3) The structure has broken into sub-structures
-            '''
-
-            #get the match and get its cluster id
-            match = list(possibleMatches)[0]
-            match_id = match.get_cluster_id()
-
-            #perform a set difference to determine gain/loss of subunits
-            old_not_new = set(match.get_body_ids()).difference(set(cluster.get_body_ids()))
-            new_not_old = set(cluster.get_body_ids()).difference(set(match.get_body_ids()))
-
-            # print(old_not_new, new_not_old)
-
-            #define a similarity, max of the two set differences
-            similarity = max(len(old_not_new), len(new_not_old))
-            # similarity = len(old_not_new)
-            # print("Similarity (Break): ", similarity)
-            # print("set diffs: ", old_not_new, new_not_old)
-
-            #if two or more particles break off
-            if len(old_not_new) > 1:
-
-                #check if this match has been assigned already
-                if match_id in label_dict: #removed .keys() in case this breaks
-
-                    #check if the current similarity score is better 
-                    if similarity < label_dict[match_id][0]:
-
-                        #TODO implement a steal id that pops the id and sets it to -1
-
-                        #current cluster is more similar to match. overwrite old cluster
-                        old_cluster = label_dict[match_id][1]
-                        old_cluster.set_cluster_id(-1)
-
-                        #set label for the new cluster
-                        #match.update(cluster)
-                        tentative_updates[match] = cluster
-                        cluster.set_cluster_id(match_id)
-                        label_dict[match_id] = [similarity, cluster]
-
-                        # print("Better match found. Updated cluster match to ", match_id)
-
-                        #add the old cluster to the queue for reassignment
-                        queue.append(old_cluster)
-
-                    #if not, assign a new cluster
-                    else: 
-
-                        #get the new cluster id
-                        cluster_num = len(cluster_info)
-
-                        #update the cluster with the new id, add entry to cluster_info
-                        cluster.set_cluster_id(cluster_num)
-                        label_dict[cluster_num] = [0, cluster]
-                        cluster_info.append(ClusterInfo(cluster, frame, observer))
-                        cluster_info[-1].set_parent(match)
-                        # print("Match found with worse similarity. Created new cluster with id ", cluster_num)
-
-                #this matching old cluster has not been assigned, so assign it
-                else:
-
-                    #update the old cluster with new. This sets the old_bodies references to cluster
-                    #match.update(cluster)
-                    tentative_updates[match] = cluster
-
-                    #grab the cluster id and set it for the new cluster
-                    cluster_id = match.get_cluster_id()
-                    cluster.set_cluster_id(cluster_id)
-                    label_dict[cluster_id] = [similarity, cluster]
-                    
-                    # print("matched cluster to existing ", cluster_id)
-
-
-            #if the difference is 1 or 0, same cluster or only lost one monomer
-            else:
-
-                #update the old cluster with new. This sets the old_bodies references to cluster
-                #match.update(cluster)
-                tentative_updates[match] = cluster
-
-                #grab the cluster id and set it for the new cluster
-                cluster_id = match.get_cluster_id()
-                cluster.set_cluster_id(cluster_id)
-                label_dict[cluster_id] = [similarity, cluster]
-                
-                # print("matched cluster to existing ", cluster_id)
-
-        else:
-
-            '''
-            If there is more than one potential match, that means that two or more clusters from the
-            previous timestep have coalesced during this frame. If one cluster is bigger than the 
-            others, keep that one. 
-            May also be able to get here if cluster merge and break at the same time. 
-            '''
-            
-            # print("Clusters have merged")
-
-            #get a list of possible matches, and assign a similarity to each
-            possible_list   = list(possibleMatches)
-            similarity_vals = []
-            for possible_match in possible_list:
-
-                #do a set difference between cluster and possible
-                old_not_new = set(possible_match.get_body_ids()).difference(set(cluster.get_body_ids()))
-                new_not_old = set(cluster.get_body_ids()).difference(set(possible_match.get_body_ids()))
-                similarity = max(len(old_not_new), len(new_not_old))
-                # similarity = len(new_not_old)
-                # print("Similarity (Merge): ", similarity)
-
-                similarity_vals.append(similarity)
-
-            #sort the possibilities by similarity
-            poss_aug = [(similarity_vals[i], i) for i in range(len(similarity_vals))]
-            poss_aug.sort()
-            sorted_sim,permutation = zip(*poss_aug)
-            sorted_poss = [possible_list[i] for i in permutation]
-
-            #loop through sorted list. assign this cluster the first un-assigned match
-            match_flag = False
-            for possibility in sorted_poss:
-
-                #check if match not already assigned. If not, assign it. 
-                match_id = possibility.get_cluster_id()
-                if (not match_flag):
-                    if (match_id not in label_dict): #removed .keys() in case things break
-
-                        # possibility.update(cluster)
-                        tentative_updates[possibility] = cluster
-
-                        #set id for the new cluster
-                        cluster.set_cluster_id(match_id)
-                        label_dict[match_id] = [similarity, cluster]
-                        match_flag = True
-                        assigned_id = match_id
-
-                    else:
-                        continue
-
-                #match is already assigned, set other possibilities to update to it and die 
-                else:
-
-                    merge_updates.append((match_id, assigned_id))
-
-            #if no match has been made after all possibilities, create new cluster
-            if not match_flag:
-
-                #i dont think its possible to get here, but we shall see
-                # print("All clusters assigned. Creating new cluster")
-
-                #get the new cluster id
-                cluster_num = len(cluster_info)
-
-                #update the cluster with the new id, add entry to cluster_info
-                cluster.set_cluster_id(cluster_num)
-                label_dict[cluster_num] = [0, cluster]
-                cluster_info.append(ClusterInfo(cluster, frame, observer))
-                # print("created new (merge) cluster with id ", cluster_num)
-
-
-    #loop over the tentative updates and apply them. do monomer checking first
-    for key in tentative_updates.keys():
-
-        #get lists of the body ids in the pre and post update clusters
-        past_ids = set(key.get_body_ids())
-        new_ids  = set(tentative_updates[key].get_body_ids())
-
-        #determine how many monomers were gained and lost from these id sets
-        m_gain, m_lost = get_monomer_stats(past_ids, new_ids, old_bodies, bodies)
-        # print("Monomer gain {}, Monomer loss {}".format(m_gain, m_lost))
-
-        #update the monomer stats
-        if m_gain > 0:
-
-            cluster_id = key.get_cluster_id()
-            monomer_frac = observer.previous_monomer
-            cluster_info[cluster_id].add_monomers(tentative_updates[key], m_gain, monomer_frac)
-
-        if m_lost > 0:
-
-            cluster_id = key.get_cluster_id()
-            cluster_info[cluster_id].remove_monomers(key, m_lost)
-
-
-        #perform the update on the cluster
-        key.update(tentative_updates[key])
-
-    #loop over the now matched clusters and update the cluster_info
-    #(can loop over all keys in the label_dict to avoid unnec work)
-    for key in list(label_dict.keys()):
-
-        #get the cluster
-        updated_cluster = label_dict[key][1]
-
-        #update the clusterinfo
-        cluster_info[key].update_data(updated_cluster, frame, observer.current_monomer)
-        # print("Updated cluster {}".format(key))
-
-    #loop over the clusters that merged and are no longer tracked
-    for update_pair in merge_updates:
-
-        old_id = update_pair[0]
-        current_id = update_pair[1]
-        current_cluster = label_dict[current_id][1]
-
-        #update to cluster, but also set a death
-        cluster_info[old_id].update_data(current_cluster, frame, observer.current_monomer)
-        cluster_info[old_id].kill(frame)
-        # print("Updated and killed cluster {} (Merge)".format(old_id))
-
-
-    return clusters, cluster_info
-
-
-def get_data_from_snap(snap, sim, frame):
-    #go from a snapshot to a frame object with bodies, clusters, and other data
-
-    #get a list of bodies to analyze
-    bodies = body.create_bodies(snap, sim)
-
-    #init a dictionary to store the bonds present - init with empty lists for each body_id
-    bond_dict = dict()
-    for bod in bodies:
-        bond_dict[bod.get_id()] = []
-
-    #determine the bond network using the list of bodies
-    body.get_bonded_bodies(bodies, sim, bond_dict)
-
-    #determine groups of bonded structures
-    G = get_groups(bond_dict)
-
-    #for each group, create a cluster
-    clusters = []
-    num_monomers = 0
-    for group in G:
-
-        #check for dimers or larger, create a cluster object
-        if len(group) > 1:
-
-            #extract the involved bodies from the group and create a cluster
-            body_list = [bodies[q] for q in group]
-            clusters.append(Cluster(body_list, frame))
-
-        #increment the number of free monomers
-        else:
-            num_monomers += 1
-
-    #set the monomer fraction
-    monomer_frac = num_monomers / len(bodies)
-    
-    #create a Frame object for this frame and return it
-    current_frame = Frame(bodies, clusters, frame, monomer_frac) 
-    return current_frame
-
-
-
-def track_clustering(snap, sim, frame, cluster_info, old_frame, observer):
+def track_clustering(snap, sim, frame_num, cluster_info, old_frame, observer):
     #compute and update cluster stats from the given frame
 
     #determine clusters and init a Frame object with the needed data
-    current_frame = get_data_from_snap(snap, sim, frame)
+    current_frame = frame.get_data_from_snap(snap, sim, frame_num)
 
     #map the clusters from previous timestep to this step to assign labels
-    print("Frame {} updates:".format(frame))
-    # clusters, cluster_info = update_clusters(clusters, cluster_info, bodies, old_bodies, 
-    #                                          frame, observer)
-
+    # print("Frame {} updates:".format(frame_num))
     current_frame.update(cluster_info, old_frame, observer)
-
-    #update the previous monomer frac
-    # observer.previous_monomer = monomer_frac
 
     #return updated cluster info and the bodies to be used as old_bodies next frame
     return cluster_info, current_frame
